pyproject.toml
[project]
name = "pms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4,<6",
    "django>=5,<6",
    "django-filter>=24.2,<25",
    "djangorestframework>=3.15,<4",
    "djangorestframework-camel-case>=1.4.2",
    "drf-nested-routers>=0.95.0",
    "drf-spectacular>=0.29.0",
    "gevent>=24.11.1",
    "gunicorn>=23.0.0",
    "psycopg[binary]>=3.3.2",
    "pyjwt>=2.11.0",
    "redis>=7.1.0",
    "rules>=3.5",
    "whitenoise>=6.11.0",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
]
